requirements.txt
faiss-cpu==1.8.0
open_clip_torch==2.24.0
torch>=2.1
torchvision>=0.16
Pillow
numpy<2.0.0
tqdm
fastapi
uvicorn
python-multipart
orjson

